# public/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Lumen: Loja de vendas",
        "version": "0.1"
    },
    "paths": {
...
        "/manufacturers/create": {
            "post": {
                "tags": [
                    "Fabricante"
                ],
                "summary": "Cadastrar novo fabricante",
                "operationId": "9970f27ecf2b9a084f7b22d7372afde7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "manufacturer_name",
                                    "manufacturer_social_name",
                                    "cnpj"
                                ],
                                "properties": {
                                    "manufacturer_name": {
                                        "type": "string",
                                        "example": "Zaeli"
                                    },
                                    "manufacturer_social_name": {
                                        "type": "string",
                                        "example": "Zaeli ltda."
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "example": "01.001.001/00001-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fabricante criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "manufacturer_name": {
                                            "type": "string",
                                            "example": "Zaeli"
                                        },
                                        "manufacturer_social_name": {
                                            "type": "string",
                                            "example": "Zaeli ltda."
                                        },
                                        "cnpj": {
                                            "type": "string",
                                            "example": "01.001.001/00001-30"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/manufacturers/{id}": {
            "get": {
                "tags": [
                    "Fabricante"
                ],
                "summary": "Buscar um fabricante",
                "operationId": "85d33ac103fcaee11e036ecd70d088d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do fabricante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aqui está o fabricante desejado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "manufacturer_name": {
                                            "type": "string",
                                            "example": "Zaeli"
                                        },
                                        "manufacturer_social_name": {
                                            "type": "string",
                                            "example": "Zaeli ltda."
                                        },
                                        "cnpj": {
                                            "type": "string",
                                            "example": "01.001.001/00001-30"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/manufacturers/update/{id}": {
            "put": {
                "tags": [
                    "Fabricante"
                ],
                "summary": "Alterar um cadastro",
                "operationId": "c001d25ca6c46548ab4ee6ea671ccfb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do fabricante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "manufacturer_name",
                                    "manufacturer_social_name",
                                    "cnpj"
                                ],
                                "properties": {
                                    "manufacturer_name": {
                                        "type": "string",
                                        "example": "Zaeli"
                                    },
                                    "manufacturer_social_name": {
                                        "type": "string",
                                        "example": "Zaeli ltda."
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "example": "01.001.001/00001-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fabricante alterado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "manufacturer_name": {
                                            "type": "string",
                                            "example": "Zaeli"
                                        },
                                        "manufacturer_social_name": {
                                            "type": "string",
                                            "example": "Zaeli ltda."
                                        },
                                        "cnpj": {
                                            "type": "string",
                                            "example": "01.001.001/00001-30"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/manufacturers/delete/{id}": {
            "delete": {
                "tags": [
                    "Fabricante"
                ],
                "summary": "Deletar um fabricante pelo ID",
                "operationId": "5c3edbf7260c6a243f30ed7f19c06dc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do fabricante a ser deletado. OBS: Se o fabricante tiver algum produto no sistema,\r\n               A deleção não será possível.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Fabricante deletado com sucesso"
                    },
                    "404": {
                        "description": "Fabricante não encontrado"
                    }
                }
            }
        },
...
    },
...
}